requirements.txt
# Core dependencies
streamlit>=1.28.0
pandas>=2.0.0
numpy>=1.24.0
openpyxl>=3.1.0

# Google APIs
google-auth>=2.23.0
google-auth-oauthlib>=1.1.0
google-auth-httplib2>=0.1.1
google-api-python-client>=2.100.0
gspread>=5.11.0

# AI/ML APIs
openai>=1.3.0
anthropic>=0.5.0
google-cloud-aiplatform>=1.36.0

# Optimization
scipy>=1.11.0
pulp>=2.7.0
scikit-learn>=1.3.0

# Data visualization
plotly>=5.17.0
matplotlib>=3.8.0
seaborn>=0.13.0

# Scheduling
schedule>=1.2.0
apscheduler>=3.10.4

# Email notifications
python-dotenv>=1.0.0
sendgrid>=6.10.0

# Utilities
python-dateutil>=2.8.2
pytz>=2023.3
requests>=2.31.0
pyyaml>=6.0.1

# Development
pytest>=7.4.0
black>=23.10.0
flake8>=6.1.0
mypy>=1.6.0
